pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-tracker"
version = "0.1.0"
description = "AI-powered arXiv paper tracking and summarization tool"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "roachbot"}
]
dependencies = [
    "httpx>=0.27.0",
    "feedparser>=6.0.10",
    "click>=8.1.7",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
paper-tracker = "paper_tracker.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["paper_tracker*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
